[chore] Switch to yarn, update packages
Move to yarn. Update all packages. Fix tests
broken by updating all packages.

Also move to using saucelabs for multi-browser
tests, so that we can catch problems more
quickly.
