+++ .travis.yml
@@ -1,14 +1,25 @@
 language: node_js
 node_js:
-- '4.3.2'
+  - node
+  - '6'
+cache:
+  yarn: true
+  directories:
+  - node_modules
+  - bower_components
 before_install:
 - export DISPLAY=:99.0
 - sh -e /etc/init.d/xvfb start
 before_script:
 - npm install -g bower gulp
-- npm install
 - bower install
 script: gulp
+addons:
+  sauce_connect: true
 notifications:
   slack:
     secure: RMNcUVDT7HTjK0Zjh/PXidyntk1zM1sOEVrsGxPjN8EGEPSXuaperQ02ckhzkxV3jUEOIw2WkFjJ0hBYvQy+h3zj57gZ0w28IiNfiUBr02v3w5iySOh/lOrbsbwO/2QVo5O9cEkppzO2eIaBGwc2FgWvwZ2mctaH0FmLOekG568=
+env:
+  global:
+  - secure: a3+b3mg3/QelYIT8bXVAxMsD9ISNkiKXx07IDG0ouG78cNX7JE/eVGM6pqBerkoANCVdLXbaSPMxLYBzfKoZEqpef5VfReC8dmzqw2ZfAKuaD/ROcJXL/iA+fgR48LPkKdwKX9aQoRk2eg8AcJsKoe7/gIYjPRsD6VLKk2s1NqY=
+  - secure: VRD2Hxm1k/WV4ISZ8mr2sBjfnaCqSpLkQUEgDXlwjfFXQvzbT4YS/I8qRCZiLBFUsckq2inFr9x1EYT8CPIxIZryj2cVwcVXxHMrMMmRxhS0qYxybID9gP/Q/obpxZBNYbHCXzJdpP8D246HhzUdQTorcz5KtvM1AIDX0cFbm7o=
