+++ package.json
@@ -20,34 +20,39 @@
   ],
   "author": "Cameron Will <scwill@us.ibm.com>",
   "license": "Apache-2.0",
-  "dependencies": {},
+  "dependencies": {
+    "gulp-webserver": "^0.9.1",
+    "jshint": "^2.9.4"
+  },
   "devDependencies": {
-    "event-stream": "3.3.0",
+    "event-stream": "3.3.4",
     "express": "4.14.1",
     "gulp": "^3.9.1",
-    "gulp-bump": "^2.0.1",
+    "gulp-bump": "^2.6.1",
     "gulp-changed": "1.3.2",
-    "gulp-concat": "^2.6.0",
+    "gulp-concat": "^2.6.1",
     "gulp-exit": "0.0.2",
-    "gulp-filter": "^3.0.1",
+    "gulp-filter": "^5.0.0",
     "gulp-footer": "1.0.5",
-    "gulp-git": "^1.6.0",
+    "gulp-git": "^1.12.0",
     "gulp-header": "1.8.8",
-    "gulp-jscs": "^3.0.2",
+    "gulp-jscs": "^4.0.0",
     "gulp-jscs-stylish": "1.4.0",
-    "gulp-jshint": "2.0.1",
-    "gulp-load-plugins": "1.2.2",
+    "gulp-jshint": "2.0.4",
+    "gulp-load-plugins": "1.4.0",
     "gulp-protractor": "3.0.0",
     "gulp-tag-version": "^1.3.0",
-    "gulp-uglify": "^2.0.0",
-    "jasmine-core": "2.5.0",
-    "jshint-stylish": "1.0.1",
-    "karma": "0.12.31",
-    "karma-chrome-launcher": "^0.2.2",
-    "karma-coverage": "0.2.7",
-    "karma-coveralls": "0.1.5",
-    "karma-jasmine": "0.3.5",
-    "protractor": "^4.0.4",
+    "gulp-uglify": "^2.0.1",
+    "gulp-util": "^3.0.8",
+    "jasmine-core": "2.5.2",
+    "jshint-stylish": "2.2.1",
+    "karma": "1.4.0",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-coverage": "1.1.1",
+    "karma-coveralls": "1.1.2",
+    "karma-jasmine": "1.1.0",
+    "karma-sauce-launcher": "^1.1.0",
+    "protractor": "^5.0.0",
     "requirejs": "2.3.2",
     "string-mask": "0.3.0"
   }
